+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Introduce `ServerConstants` and move several constants from `SharedConstants` to `ServerConstants` to align with changes applied in version 6.
 * Rename `getID()` to `getId()` and `getTypeID()` to `getTypeId()` to align with version 6.
+* Expose `ChannelMetaData.isExternallyVisible()` to align with version 6.
 
 ### [v5.101](https://github.com/replicant4j/replicant/tree/v5.101) (2019-12-02) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v5.100...v5.101)